chore: Move NewDeviceSystemMessageCell into its own file
