+++ wire-ios/Wire-iOS/Sources/UserInterface/Conversation/Content/Cells/ConfigurationMessageCell/Components/NewDeviceSystemMessageCell.swift
@@ -1,0 +1,88 @@
+//
+// Wire
+// Copyright (C) 2024 Wire Swiss GmbH
+//
+// This program is free software: you can redistribute it and/or modify
+// it under the terms of the GNU General Public License as published by
+// the Free Software Foundation, either version 3 of the License, or
+// (at your option) any later version.
+//
+// This program is distributed in the hope that it will be useful,
+// but WITHOUT ANY WARRANTY; without even the implied warranty of
+// MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE. See the
+// GNU General Public License for more details.
+//
+// You should have received a copy of the GNU General Public License
+// along with this program. If not, see http://www.gnu.org/licenses/.
+//
+
+import Foundation
+import UIKit
+import WireDataModel
+
+class NewDeviceSystemMessageCell: ConversationIconBasedCell, ConversationMessageCell {
+
+    static let userClientURL: URL = URL(string: "settings://user-client")!
+
+    var linkTarget: LinkTarget?
+
+    enum LinkTarget {
+        case user(UserType)
+        case conversation(ZMConversation)
+    }
+
+    struct Configuration {
+        let attributedText: NSAttributedString?
+        var icon: UIImage?
+        var linkTarget: LinkTarget
+    }
+
+    override init(frame: CGRect) {
+        super.init(frame: frame)
+
+        setupView()
+    }
+
+    @available(*, unavailable)
+    required init?(coder aDecoder: NSCoder) {
+        fatalError("init?(coder aDecoder: NSCoder) is not implemented")
+    }
+
+    func setupView() {
+        lineView.isHidden = false
+    }
+
+    func configure(with object: Configuration, animated: Bool) {
+        attributedText = object.attributedText
+        imageView.image = object.icon
+        linkTarget = object.linkTarget
+    }
+
+}
+
+// MARK: - UITextViewDelegate
+
+extension NewDeviceSystemMessageCell {
+
+    public override func textView(
+        _ textView: UITextView,
+        shouldInteractWith url: URL,
+        in characterRange: NSRange,
+        interaction: UITextItemInteraction
+    ) -> Bool {
+
+        guard let linkTarget = linkTarget,
+              url == type(of: self).userClientURL,
+              let zClientViewController = ZClientViewController.shared else { return false }
+
+        switch linkTarget {
+        case .user(let user):
+            zClientViewController.openClientListScreen(for: user)
+        case .conversation(let conversation):
+            zClientViewController.openDetailScreen(for: conversation)
+        }
+
+        return false
+    }
+
+}
